Move format_date_range to cli/utils/_date_formatter.py
