+++ git_dev_metrics/cli/wizards/dashboard_wizard.py
@@ -4,8 +4,8 @@ from pathlib import Path
 import typer
 
 from ...metrics.loader import load_snapshot_for_months
-from ...utils.date_utils import format_date_range
 from ..runners.dashboard_runner import write_and_open_dashboard
+from ..utils._date_formatter import format_date_range
 from ._wizard import _prompt_months, pick_months
 
 YearMonth = tuple[int, int]
